backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
# Pillow==10.0.1
# numpy==1.24.3
# scipy==1.11.4
